+++ lib/lindi.dart
@@ -3,4 +3,3 @@ library;
 export 'src/core/lindi_view_model.dart';
 export 'src/core/lindi_injector.dart';
 export 'src/builders/lindi_builder.dart';
-export 'src/builders/lindi_multi_builder.dart';