+++ lib/src/builders/lindi_builder.dart
@@ -1,20 +1,25 @@
 import 'package:flutter/material.dart';
 import 'package:lindi/lindi.dart';
 
-/// A widget that listens to a `LindiViewModel` and rebuilds
-/// whenever the `viewModel` changes, calling the provided builder function.
+/// A widget that listens to a list of `LindiViewModel`s and rebuilds
+/// whenever any of the view models change, calling the provided builder function.
 class LindiBuilder extends StatefulWidget {
-  /// The view model that this widget listens to for changes.
-  final LindiViewModel viewModel;
+  /// A list of view models that this widget listens to for changes.
+  final List<LindiViewModel> viewModels;
+
+  /// Callback function executed when the `viewModel` updates.
+  final void Function(BuildContext context, LindiViewModel viewModel)? listener;
 
   /// The builder function that defines how the UI should be built.
   final Widget Function(BuildContext context) builder;
 
-  const LindiBuilder({
+  LindiBuilder({
     super.key,
-    required this.viewModel,
+    required this.viewModels,
+    this.listener,
     required this.builder,
-  });
+  }) : assert(viewModels.isNotEmpty,
+            "LindiBuilder requires at least one LindiViewModel");
 
   @override
   State<LindiBuilder> createState() => _LindiBuilderState();
@@ -24,32 +29,48 @@ class _LindiBuilderState extends State<LindiBuilder> {
   @override
   void initState() {
     super.initState();
-    // Adds a listener to the view model to trigger a rebuild when it changes.
-    widget.viewModel.addListener(_onViewModelChanged);
+    // Adds listeners to each view model in the list to trigger a rebuild
+    // when any of them change.
+    _addListeners();
+  }
+
+  void _addListeners() {
+    for (var viewModel in widget.viewModels) {
+      viewModel.addListener(_onViewModelChanged);
+    }
+  }
+
+  void _removeListeners() {
+    for (var viewModel in widget.viewModels) {
+      viewModel.removeListener(_onViewModelChanged);
+    }
   }
 
   @override
   void didUpdateWidget(covariant LindiBuilder oldWidget) {
     super.didUpdateWidget(oldWidget);
 
-    // If the view model changes, remove the listener from the old view model
-    // and add it to the new one.
-    if (oldWidget.viewModel != widget.viewModel) {
-      oldWidget.viewModel.removeListener(_onViewModelChanged);
-      widget.viewModel.addListener(_onViewModelChanged);
+    // If the list of view models has changed, remove listeners from
+    // the old view models and add them to the new ones.
+    if (oldWidget.viewModels != widget.viewModels) {
+      _removeListeners();
+      _addListeners();
     }
   }
 
   @override
   void dispose() {
-    // Removes the listener when the widget is disposed to prevent memory leaks.
-    widget.viewModel.removeListener(_onViewModelChanged);
+    // Removes listeners from all view models when the widget is disposed
+    // to prevent memory leaks.
+    _removeListeners();
     super.dispose();
   }
 
-  /// Callback triggered when the `viewModel` changes.
-  /// This forces the widget to rebuild by calling `setState`.
-  void _onViewModelChanged() {
+  /// Callback triggered when any of the `viewModels` changes.
+  void _onViewModelChanged(LindiViewModel viewModel) {
+    if (widget.listener != null) {
+      widget.listener!(context, viewModel);
+    }
     setState(() {});
   }