+++ test/lindi_builder_test.dart
@@ -28,24 +28,50 @@ class AsyncTestLindiViewModel extends LindiViewModel<String, String> {
   }
 
   void testError() {
-    setData('error');
+    setError('error');
   }
 }
 
 void main() {
   group('LindiBuilder', () {
-    testWidgets(
-        'should rebuild when CounterTestLindiViewModel updates with notify()',
-        (tester) async {
-      final viewModel = CounterTestLindiViewModel();
+    testWidgets('should rebuild when LindiViewModels updates', (tester) async {
+      final counterViewModel = CounterTestLindiViewModel();
+      final asyncViewModel = AsyncTestLindiViewModel();
+
+      bool listenerLoadingCalled = false;
+      bool listenerDataCalled = false;
+      bool listenerErrorCalled = false;
 
       await tester.pumpWidget(
         MaterialApp(
           home: LindiBuilder(
-            viewModel: viewModel,
+            viewModels: [counterViewModel, asyncViewModel],
+            listener: (context, viewModel) {
+              if (asyncViewModel.isLoading) {
+                listenerLoadingCalled = true;
+              }
+              if (asyncViewModel.hasError) {
+                listenerErrorCalled = true;
+              }
+              if (asyncViewModel.hasData) {
+                listenerDataCalled = true;
+              }
+            },
             builder: (context) {
-              return Text('${viewModel.counter}',
-                  textDirection: TextDirection.ltr);
+              return Column(
+                children: [
+                  Text('${counterViewModel.counter}',
+                      textDirection: TextDirection.ltr),
+                  if (asyncViewModel.isLoading)
+                    Text('loading', textDirection: TextDirection.ltr),
+                  if (asyncViewModel.hasError)
+                    Text('${asyncViewModel.error}',
+                        textDirection: TextDirection.ltr),
+                  if (asyncViewModel.hasData)
+                    Text('${asyncViewModel.data}',
+                        textDirection: TextDirection.ltr),
+                ],
+              );
             },
           ),
         ),
@@ -53,59 +79,36 @@ void main() {
 
       // Initial state should be 0
       expect(find.text('0'), findsOneWidget);
+      expect(listenerLoadingCalled, false);
 
       // Update view model state
-      viewModel.increment();
+      counterViewModel.increment();
+      asyncViewModel.testLoading();
       await tester.pump();
 
       // UI should update to reflect new state
       expect(find.text('1'), findsOneWidget);
+      expect(find.text('loading'), findsOneWidget);
+      expect(listenerLoadingCalled, true);
+      expect(listenerErrorCalled, false);
 
       // Update view model state
-      viewModel.decrement();
-      viewModel.decrement();
+      counterViewModel.decrement();
+      counterViewModel.decrement();
+      asyncViewModel.testError();
       await tester.pump();
 
       // UI should update to reflect new state
       expect(find.text('-1'), findsOneWidget);
-    });
-
-    testWidgets(
-        'should rebuild when AsyncTestLindiViewModel updates with setLoading, setData, setError',
-        (tester) async {
-      final viewModel = AsyncTestLindiViewModel()..testLoading();
-
-      await tester.pumpWidget(
-        MaterialApp(
-          home: LindiBuilder(
-            viewModel: viewModel,
-            builder: (context) {
-              if (viewModel.isLoading) {
-                return Text('loading', textDirection: TextDirection.ltr);
-              }
-              if (viewModel.hasError) {
-                return Text(viewModel.error!, textDirection: TextDirection.ltr);
-              }
-              return Text(viewModel.data!, textDirection: TextDirection.ltr);
-            },
-          ),
-        ),
-      );
-
-      // Initial state should be 0
-      expect(find.text('loading'), findsOneWidget);
-
-      // Update view model state
-      viewModel.testData();
-      await tester.pump();
-      // UI should update to reflect new state
-      expect(find.text('success'), findsOneWidget);
-
-      // Update view model state
-      viewModel.testError();
-      await tester.pump();
-      // UI should update to reflect new state
       expect(find.text('error'), findsOneWidget);
+      expect(listenerErrorCalled, true);
+      expect(listenerDataCalled, false);
+
+      asyncViewModel.testData();
+      await tester.pump();
+
+      expect(find.text('success'), findsOneWidget);
+      expect(listenerDataCalled, true);
     });
   });
 }